requirements.txt
Django>=1.7,<1.10
django-extensions
Pillow >=3.2.0
whitenoise
django-tastypie
